+++ test/test_helper.exs
@@ -1,1 +1,4 @@
 ExUnit.start()
+{:ok, _ } = :application.ensure_all_started(:httparrot)
+{:ok, _} = :application.ensure_all_started(:proxy)
+Task.start(fn -> Proxy.start([]) end)